GENAI/IN226108102_GENAI/RAG Based Customer Support Assistant/requirements.txt
langchain>=0.1.0
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
langchain-chroma
chromadb
langgraph
pypdf
python-dotenv
fpdf
streamlit
torch
torchvision
